evals/src/make_traces.py
#!/usr/bin/env python3
"""
EXAID Evaluation - Trace Generation Script (Stub)

This is a PLACEHOLDER script that will later integrate with MAC (Multi-Agent Collaboration)
to generate MAS traces from clinical cases.

Current behavior:
- Creates stub trace files in the expected format
- Documents the integration points for MAC

MAC Integration Points:
- Replace the stub_generate_traces() function with MAC trace generation
- MAC will provide the multi-agent reasoning traces
- Output format follows exaid.trace.schema.json

Usage:
    python make_traces.py --config configs/mas_generation.yaml
    python make_traces.py --cases data/cases/ --output data/traces/
"""

import argparse
import gzip
import json
import sys
from datetime import datetime, timezone
from pathlib import Path
from typing import Iterator

import yaml


def load_config(config_path: Path) -> dict:
    """Load MAS generation configuration."""
    with open(config_path, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)


def generate_trace_id(case_id: str, sequence: int) -> str:
    """Generate a unique trace ID."""
    return f"trc-{case_id.replace('case-', '')}-{sequence:03d}"


def stub_generate_traces(case_id: str, case_content: dict) -> Iterator[dict]:
    """
    STUB: Generate placeholder traces for a case.
    
    TODO: Replace this function with MAC integration
    
    MAC Integration Notes:
    - MAC will process the clinical case through multiple agents
    - Each agent's reasoning output becomes a trace record
    - Traces should be yielded in temporal order
    - Token counts should be computed using the appropriate tokenizer
    
    Args:
        case_id: Unique case identifier
        case_content: Clinical case data
        
    Yields:
        Trace records conforming to exaid.trace.schema.json
    """
    # Placeholder: Generate stub traces
    # In real implementation, MAC would generate these
    
    agents = ["orchestrator", "cardiology_agent", "laboratory_agent"]
    stub_contents = [
        "Analyzing clinical presentation and initial assessment...",
        "Evaluating cardiac biomarkers and ECG findings...",
        "Reviewing laboratory results and metabolic panel...",
    ]
    
    timestamp = datetime.now(timezone.utc)
    
    for seq, (agent, content) in enumerate(zip(agents, stub_contents)):
        trace = {
            "schema_name": "exaid.trace",
            "schema_version": "1.0.0",
            "trace_id": generate_trace_id(case_id, seq),
            "case_id": case_id,
            "agent_id": agent,
            "sequence_num": seq,
            "timestamp": timestamp.isoformat(),
            "content": f"[STUB] {content}",
            "token_count": len(content.split()),  # Approximate
            "metadata": {
                "model": "stub",
                "temperature": 0.0,
                "latency_ms": 0
            }
        }
        yield trace


def load_cases(cases_dir: Path) -> Iterator[tuple[str, dict]]:
    """
    Load clinical cases from the cases directory.
    
    Args:
        cases_dir: Path to cases directory
        
    Yields:
        Tuples of (case_id, case_content)
    """
    if not cases_dir.exists():
        print(f"Cases directory does not exist: {cases_dir}")
        print("Creating stub case for demonstration...")
        # Generate a stub case
        yield "case-stub-001", {
            "patient_id": "stub-001",
            "chief_complaint": "Stub clinical case for scaffold testing",
            "content": "This is a placeholder case."
        }
        return
    
    for case_file in sorted(cases_dir.glob("*.json")):
        case_id = f"case-{case_file.stem}"
        with open(case_file, "r", encoding="utf-8") as f:
            case_content = json.load(f)
        yield case_id, case_content


def write_traces(
    traces: Iterator[dict],
    output_path: Path,
    compress: bool = True
) -> int:
    """
    Write traces to a JSONL file.
    
    Args:
        traces: Iterator of trace records
        output_path: Output file path
        compress: Whether to gzip compress the output
        
    Returns:
        Number of traces written
    """
    output_path.parent.mkdir(parents=True, exist_ok=True)
    
    count = 0
    open_func = gzip.open if compress else open
    mode = "wt" if compress else "w"
    
    with open_func(output_path, mode, encoding="utf-8") as f:
        for trace in traces:
            f.write(json.dumps(trace) + "\n")
            count += 1
    
    return count


def main():
    parser = argparse.ArgumentParser(
        description="Generate MAS traces from clinical cases (stub implementation)"
    )
    parser.add_argument(
        "--config",
        type=Path,
        default=Path("configs/mas_generation.yaml"),
        help="MAS generation configuration file"
    )
    parser.add_argument(
        "--cases",
        type=Path,
        default=Path("data/cases"),
        help="Input cases directory"
    )
    parser.add_argument(
        "--output",
        type=Path,
        default=Path("data/traces"),
        help="Output traces directory"
    )
    parser.add_argument(
        "--no-compress",
        action="store_true",
        help="Disable gzip compression"
    )
    
    args = parser.parse_args()
    
    print("=" * 60)
    print("EXAID Trace Generation (STUB)")
    print("=" * 60)
    print()
    print("NOTE: This is a STUB implementation.")
    print("MAC integration will replace the trace generation logic.")
    print()
    
    # Load configuration if it exists
    config = {}
    if args.config.exists():
        config = load_config(args.config)
        print(f"Loaded config: {args.config}")
        print(f"  Mode: {config.get('mode', 'stub')}")
        print()
    
    # Check for MAC integration
    if config.get("mac", {}).get("enabled", False):
        print("ERROR: MAC integration is enabled but not yet implemented.")
        print("Set mac.enabled: false in config or implement MAC integration.")
        return 1
    
    # Process cases
    total_traces = 0
    total_cases = 0
    
    for case_id, case_content in load_cases(args.cases):
        total_cases += 1
        print(f"Processing: {case_id}")
        
        # Generate traces (stub)
        traces = stub_generate_traces(case_id, case_content)
        
        # Determine output path
        output_file = args.output / f"{case_id}.jsonl"
        if not args.no_compress:
            output_file = output_file.with_suffix(".jsonl.gz")
        
        # Write traces
        count = write_traces(traces, output_file, compress=not args.no_compress)
        total_traces += count
        print(f"  Written: {count} traces -> {output_file}")
    
    print()
    print("=" * 60)
    print(f"COMPLETE: {total_cases} cases, {total_traces} traces")
    print("=" * 60)
    print()
    print("Next steps:")
    print("1. Integrate MAC for real trace generation")
    print("2. Run: python src/run_variants.py")
    
    return 0


if __name__ == "__main__":
    sys.exit(main())


