evals/requirements-evals.txt
# EXAID Evaluation Module - Pinned Dependencies
# Python 3.12 compatible

# Core infra
pydantic>=2.0.0,<3.0.0
jsonschema>=4.20.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
tqdm>=4.66.0,<5.0.0

# Numeric stack (wheel-available on Py3.12)
numpy>=1.26.0,<2.0.0
scipy>=1.11.0,<2.0.0
pandas>=2.1.0,<3.0.0
matplotlib>=3.8.0,<4.0.0

# NLP
spacy==3.7.4
